requirements.txt
addict==2.4.0
albumentations==1.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colorlog==6.6.0
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
easydev==0.12.0
efficientnet-pytorch==0.7.1
einops==0.4.1
entrypoints==0.4
fastai==2.7.7
fastcore==1.5.6
fastdownload==0.0.7
fastjsonschema==2.16.1
fastprogress==1.0.2
filelock==3.8.0
fitter==1.4.1
fonttools==4.34.4
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.19.5
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaggle==1.5.12
kiwisolver==1.4.4
kmeans1d==0.3.1
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mmcv==1.6.1
munch==2.5.0
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.12
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
preshed==3.0.6
pretrainedmodels==0.7.4
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1637806658031/work
seaborn==0.11.2
segmentation-models-pytorch==0.3.0
Send2Trash==1.8.0
sentry-sdk==1.8.0
setproctitle==1.2.3
shortuuid==1.0.9
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
spams==2.6.1
srsly==2.4.4
staintools==2.1.2
terminado==0.15.0
text-unidecode==1.3
thinc==8.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tinycss2==1.1.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchcontrib==0.0.2
torchstain==1.2.0
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typer==0.4.2
typing_extensions==4.1.1
urllib3==1.26.10
wandb==0.12.21
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
yacs==0.1.8
yapf==0.32.0
zipp==3.8.1
